test: refactor Changelog tests as it delegates formatting
