+++ package.json
@@ -13,6 +13,11 @@
   "jest": {
     "transform": {
       "^.+\\.tsx?$": "ts-jest"
+    },
+    "globals": {
+      "ts-jest": {
+        "diagnostics": false
+      }
     }
   },
   "devDependencies": {
